Create string-int-example3.go
Here our input is provided in the form of “id:00211” where you have the prefix “id:” and the number is fixed 5 digits, padded with zeros if shorter, this is very quickly parsable. That is why we got 211 in the output.

With this Sscanf() method, we must specify the format string to guide the parsing of the values.